Remove custom scheduling implementation and migrate to CronScheduler.AspNetCore
