+++ src/Promitor.Scraper.Host/Promitor.Scraper.Host.csproj
@@ -36,7 +36,6 @@
     <PackageReference Include="CronScheduler.AspNetCore" Version="1.0.11" />
     <PackageReference Include="Microsoft.AspNetCore.App" Version="2.2.3" />
     <PackageReference Include="Prometheus.Client.AspNetCore" Version="2.1.0" />
-    <PackageReference Include="Shuttle.Core.Cron" Version="10.0.6" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
     <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="4.0.1" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
