--- src/Promitor.Scraper.Host/Scheduling/Infrastructure/Extensions/SchedulerExtensions.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Promitor.Scraper.Host.Scheduling.Interfaces;
-
-/*
- * Based on example by Maarten Balliauw - https://blog.maartenballiauw.be/post/2017/08/01/building-a-scheduled-cache-updater-in-aspnet-core-2.html
- * Thank you!
- */
-namespace Promitor.Scraper.Host.Scheduling.Infrastructure.Extensions
-{
-    public static class SchedulerExtensions
-    {
-        public static IServiceCollection AddScheduler(this IServiceCollection services, EventHandler<UnobservedTaskExceptionEventArgs> unobservedTaskExceptionHandler)
-        {
-            return services.AddSingleton<IHostedService, SchedulerHostedService>(serviceProvider =>
-            {
-                var instance = new SchedulerHostedService(serviceProvider.GetServices<IScheduledTask>());
-                instance.UnobservedTaskException += unobservedTaskExceptionHandler;
-                return instance;
-            });
-        }
-    }
-}
