--- src/Promitor.Scraper.Host/Scheduling/Interfaces/IScheduledTask.cs
@@ -1,15 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-
-/*
- * Based on example by Maarten Balliauw - https://blog.maartenballiauw.be/post/2017/08/01/building-a-scheduled-cache-updater-in-aspnet-core-2.html
- * Thank you!
- */
-namespace Promitor.Scraper.Host.Scheduling.Interfaces
-{
-    public interface IScheduledTask
-    {
-        string Schedule { get; }
-        Task ExecuteAsync(CancellationToken cancellationToken);
-    }
-}
